% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_point.R
\name{anchor_point}
\alias{anchor_point}
\title{Anchor Point Test}
\usage{
anchor_point(
  df,
  total_observations,
  dr_estimate,
  central_tendency,
  upper_green = 1.2,
  upper_red = 1.3,
  lower_green = 0.8,
  lower_red = 0.7
)
}
\arguments{
\item{df}{A data.frame (tibble)}

\item{total_observations}{The column name of total observations variable. (character)}

\item{dr_estimate}{The column name of bad estimate variable. (character)}

\item{central_tendency}{A central tendency needs to be specified for testing.}

\item{upper_green}{(Default is 1.2)}

\item{upper_red}{(Default is 1.3)}

\item{lower_green}{(Default is 0.8)}

\item{lower_red}{(Default is 0.7)}
}
\description{
Anchor point test
}
