% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_bins.R
\name{apply_bins}
\alias{apply_bins}
\title{Apply bins}
\usage{
apply_bins(model, df, y_target)
}
\arguments{
\item{model}{ctree model object}

\item{df}{A data.frame (tibble) with independent variable.}
}
\description{
apply_bins is for applying bins rules.
}
