% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_test.R
\name{binomial_test}
\alias{binomial_test}
\title{Binomial Test}
\usage{
binomial_test(
  df,
  total_observations,
  dr_observation,
  dr_estimate,
  confidence_level = 0.95,
  tail = "one"
)
}
\arguments{
\item{df}{A data.frame (tibble)}

\item{total_observations}{The column name of total observations variable. (character)}

\item{dr_observation}{The column name of bad observation rate variable. (character)}

\item{dr_estimate}{The column name of bad estimate rate variable. (character)}

\item{confidence_level}{Confidence level (Default is 0.95)}

\item{tail}{Test tail specification, could be one or two. (Default is one)}
}
\description{
Binomial test
}
