% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminatory.R
\name{discriminatory_tests}
\alias{discriminatory_tests}
\title{Discriminatory tests}
\usage{
discriminatory_tests(
  df,
  total_observations,
  dr_observation,
  dr_estimate,
  simplfy = T
)
}
\arguments{
\item{df}{A data.frame (tibble)}

\item{total_observations}{The column name of total observations variable. (character)}

\item{dr_observation}{The column name of bad observation rate variable. (character)}

\item{dr_estimate}{The column name of bad estimate rate variable. (character)}

\item{simplfy}{(logical, Default is TRUE)}
}
\description{
Discriminatory tests (gini and ks)
}
