% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frequency_table.R
\name{get_frequency_table}
\alias{get_frequency_table}
\title{Calculate Frequency Table}
\usage{
get_frequency_table(df, x_target = NULL, y_target = NULL)
}
\arguments{
\item{df}{A data.frame (tibble) with independent variables and target variable.}

\item{x_target}{Names of independent variables. (Default is NULL)}

\item{y_target}{The name of target variable. (Default is NULL)}
}
\description{
Calculate frequency table
}
\details{
\itemize{
 \item {if y_target is not NULL, it can be factor or character.}
 \item {if y_target is not NULL, split_bins() are apllied the following variables: integer, double or numeric.}
 \item {if y_target is NULL, only factor or character variables are considered.}
 }
}
