% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_risk_table.R
\name{get_risk_table}
\alias{get_risk_table}
\title{Calculate Risk Table}
\usage{
get_risk_table(df_master, df_new, y_target, x_target = NULL, default = NULL)
}
\arguments{
\item{df_master}{A master data.frame (tibble) with independent variables and target variable.}

\item{df_new}{A new data.frame (tibble) with independent variables and target variable.}

\item{y_target}{The name of target variable.}

\item{x_target}{Names of independent variables. (Default is NULL)}

\item{default}{Default flag in the target variable (Default is NULL)}
}
\description{
Calculate risk table
}
