% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herfindahl_hirschman_index.R
\name{herfindahl_hirschman_index}
\alias{herfindahl_hirschman_index}
\title{Herfindahl Hirschman Index}
\usage{
herfindahl_hirschman_index(df, total_observations, trace = F, simplfy = T)
}
\arguments{
\item{df}{A data.frame (tibble)}

\item{total_observations}{The column name of total observations variable. (character)}

\item{trace}{show total HHI score (logical, Default is FALSE)}

\item{simplfy}{if is TRUE, return only tibble,
otherwise return list (logical, Default is TRUE)}
}
\description{
Herfindahl Hirschman Index
}
