% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herfindahl_hirschman_index.R
\name{herfindahl_hirschman_index_table}
\alias{herfindahl_hirschman_index_table}
\title{Herfindahl Hirschman Index Table}
\usage{
herfindahl_hirschman_index_table(
  df,
  df_column = "DATA",
  total_observations,
  simplfy = T
)
}
\arguments{
\item{df}{Frequency tables for variables (nested tibble)}

\item{df_column}{Nested tibble column name (Default is DATA)}

\item{total_observations}{The column name of total observations variable. (character)}

\item{simplfy}{if is TRUE, return only HHI value, otherwise detail.}
}
\description{
Herfindahl Hirschman Index table
}
