% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{psi}
\alias{psi}
\title{Population Stability Index}
\usage{
psi(df, count_observed, count_estimated, trace = F, simplfy = T)
}
\arguments{
\item{df}{A data.frame (tibble)}

\item{count_observed}{The column name of count observation variable. (character)}

\item{count_estimated}{The column name of count estimate variable. (character)}

\item{trace}{show total PSI index (logical, Default is FALSE)}

\item{simplfy}{if is TRUE, return only tibble,
otherwise return list (logical, Default is TRUE)}
}
\description{
Population stability index (PSI)
}
