% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_bins.R
\name{split_bins}
\alias{split_bins}
\title{Split bins}
\usage{
split_bins(df, y_target, p = 0.05, simplfy = T)
}
\arguments{
\item{df}{A data.frame (tibble) with independent and target variable.}

\item{y_target}{The name of target variable. (factor or character)}

\item{p}{probability for minimum bucket (Default is 0.05)}

\item{simplfy}{(Default is TRUE)}
}
\description{
split_bins is for binning using breaks.
}
