% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_light.R
\name{traffic_light_tests}
\alias{traffic_light_tests}
\title{Traffic Light Tests}
\usage{
traffic_light_tests(
  df,
  total_observations,
  dr_observation,
  dr_estimate,
  rho,
  simplfy = T
)
}
\arguments{
\item{df}{A data.frame (tibble)}

\item{total_observations}{The column name of total observations variable. (character)}

\item{dr_observation}{The column name of bad observation rate variable. (character)}

\item{dr_estimate}{The column name of bad estimate rate variable. (character)}

\item{rho}{test correlation (numeric)}

\item{simplfy}{(logical, Default is TRUE)}
}
\description{
p-value based traffic light tests (binom, normal, and correlated tests)
}
